project-asset/requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
asgiref==3.2.7
bcrypt==3.1.7
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.9.2
distlib==0.3.0
distro==1.4.0
Django==3.0.6
django-cors-headers==3.3.0
django-mysql==3.7.1
django-redis==4.12.1
djangorestframework==3.11.0
dnspython==1.16.0
dwebsocket==0.5.12
eventlet==0.30.0
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
importlib-metadata==2.0.0
install==1.3.4
ipaddr==2.2.0
IPy==1.0
lockfile==0.12.2
msgpack==0.6.2
mysqlclient==1.4.6
packaging==20.3
paramiko==2.7.1
pep517==0.8.2
progress==1.5
psutil==5.7.2
pycparser==2.20
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
tzlocal==2.1
urllib3==1.25.8
uWSGI==2.0.19.1
webencodings==0.5.1
xlrd==1.2.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
gunicorn==20.0.4
